+++ Revo.Infrastructure/Jobs/InMemory/IInMemoryJobScheduler.cs
@@ -1,0 +1,6 @@
+﻿namespace Revo.Infrastructure.Jobs.InMemory
+{
+    public interface IInMemoryJobScheduler : IJobScheduler
+    {
+    }
+}