+++ Revo.Infrastructure/Jobs/InMemory/IInMemoryJobSchedulerProcess.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Revo.Infrastructure.Jobs.InMemory
+{
+    public interface IInMemoryJobSchedulerProcess
+    {
+        void ScheduleJob(IJob job, DateTimeOffset enqueueAt, Action<IJob, Exception> errorHandler);
+    }
+}