+++ Revo.Infrastructure/Jobs/InMemory/IInMemoryJobWorkerProcess.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Revo.Infrastructure.Jobs.InMemory
+{
+    public interface IInMemoryJobWorkerProcess
+    {
+        void EnqueueJob(IJob job, Action<IJob, Exception> errorHandler);
+    }
+}